Changing place for table classes
