+++ BikeData/src/main/scala/by/itechart/internship/entities/Station.scala
@@ -1,8 +1,22 @@
 package by.itechart.internship.entities
 
+import slick.jdbc.PostgresProfile.api._
+
 case class Station(
                     station_id: Long,
                     name_station: String,
                     latitude: Double,
                     longitude: Double
                   )
+
+class StationTable(tag: Tag) extends Table[Station](tag, "station") {
+  def stationId = column[Long]("station_id", O.PrimaryKey)
+
+  def nameStation = column[String]("name_station")
+
+  def latitude = column[Double]("latitude")
+
+  def longitude = column[Double]("longitude")
+
+  def * = (stationId, nameStation, latitude, longitude).mapTo[Station]
+}
